msm8226-common: Switch to zImage-dtb
Testing changes to the device tree using fastboot boot is not possible
if it's kept separate from the zImage. The new kernel boots using the
old device tree blob. This doesn't happen when the dtb is appended to
the zImage, so switch to zImage-dtb. Moreover, images that have a
kernel with appended dtb appear to be slightly smaller than images
where the dt is kept separate.

Change-Id: Iac3038a9a2ff6344cbdcd9e4a89a4e95dbda6dfa
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>