Revert "msm8226-common: Build with SDCLANG"
I don't really like that official builds are made with some unknown
compiler not included in the synced source tree. Moreover, it's still
to be verified properly whether SD clang is the best compiler for us.
Personally, I couldn't notice major differences on a day-to-day usage,
so just go back to the easily accessible and well tested compiler.

This reverts commit 2a7980e2d5fc5ebc37eb6cef5886a46d601c773f.

Change-Id: I417bd51e5d0985f01f279e83c5b623aac723e33c
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>