sensors: falcon-peregrine: Query compass accuracy on init
The kernel uses abs events for the sensor accuracy. This means that
the event is only sent when the accuracy changes and since it doesn't
change very frequently, we might report an incorrect accuracy for a
considerable amount of time. Query the current sensor accuracy on
initialization to ensure that kernel and userspace are in sync.

Change-Id: I6580cd9ef841dbb7c31eeb16ccd32969ae2192d9